Fix broken Secure Cell tests
Remember that 0xDEADBEEF constant which got changed to 0x1337BEEF
because 32-bit systems could not handle that value? Here we are again.

This value is used as a 'corrupted' message length value. Eventually it
will be used to allocate appropriate amount of memory so that the test
can fail successfully.

Real 32-bit Linux systems can allocate 0x1337BEEF bytes of memory just
fine. It is possible due to the way memory allocation works on Linux
with glibc. However, Wasm runtime has much more strict memory limits
and is not able to reserve the memory with lazy mapping the way Linux
does it. Furthermore, instead of returning NULL from malloc() the
runtime aborts by default.

Lower the value of this constant again. 16 KB should be big enough to
be considered invalid, but small enough to be successfully allocated.