Run Emscripten tests on CirleCI
In order to compile Wasm on CircleCI we need to install Emscripten
toolchain. Do that in accordance with the official docs. After that
the build should just work out of the box. Take care to cache as much
as possible so that we don't have to download LLVM and recompile the
C standard library each build.