Explicitly mark public API of Themis
Introduce a new header <themis/themis_api.h> which defines a function
attribute marker THEMIS_API. It is used to explicitly mark API that is
considered public. These functions will be exported from the library.

Themis does not really provide separate public header files so it's not
immediately obvious which functions have to be marked public. We go with
functions visible when including <themis/themis.h>.

Currently we support only UNIX-like systems so we can limit ourselves
to just Clang and GCC. MSVC requires slightly different syntax but we'll
get back to that when we start supporting Windows again.