Initial "wasm-themis" npm package
We're going to publish WasmThemis via npm, create a package for that.
Let's be nice and provide README and LICENSE files, fill in some
informational fields in in package.json, etc. Note also these carefully
crafted .gitignore and .npmignore files. They are meant to ensure that
developers could do their thing in the working copy without accidentally
committing unnecessary files to git or publishing them on npm.

The package is currently marked as _private_ to prevent accidental
publish. However, it is going to be named "wasm-themis" and that's the
name used in README.

As for the wrapper code itself, let's start with secure key generation.
This gives us a good start as we need to provide all the scaffolding
around Node.js modules and Themis error handling. The interface stays
similar to JsThemis, but there are no compatibility requirements. Thus
we are able to provide strongly typed PrivateKey and PublicKey wrappers
over Uint8Array, and expose then as read-only properties of KeyPair.