Windows support in API macros
Add [experimental and untested] support for Windows to the API macros.
Note that Windows uses slightly different approach to DLL linkage. Just
like on Linux and macOS, exported functions should be marked on Windows.
However, _imported_ functions should be explicitly marked as well.

Effectively, this means that we should mark exported functions with
 __declspec(dllexport) when compiling DLL object files, but users of
the DLL should see functions marked with __declspec(dllimport).
We achieve that by using a special macro like THEMIS_EXPORT which is
defined only when compling Themis source code, but not when using it.