Correct install names on macOS
Actually, with this changes the install names require more attention.
We still need to use "install_name_tool", but let's apply it only
during actual installation. That way we keep the libraries in build
directory unchanged, and we can still link against them for testing.
If we change the install name during build then the libraries become
unusable unless installed to the specified paths.

Pay attention to the paths we use (build path, install path, destdir
path), and the fact that we need to fixup Soter search path in Themis
library.