Link tests dynamically: Soter
Switch to dynamic linkage for test binaries. This will allow us to
better track missing exported symbols and make sure that all public API
is actually tested.

The switch itself is more or less trivial: build the dynamic library
as the prerequisite, and link using '-Lbuild -lsoter' instead of adding
the static library to the object list. Since now the cryptographic
backend is managed by the dynamic library, we no longer have to link
against it explicitly.

Caveat for Linux: we need to set "rpath" to the build directory so that
the dynamic loader can locate Soter library there. Otherwise we'll have
to set LD_LIBRARY_PATH when running tests.