Disallow unresolved symbols during builds
The default behavior of linkers is to allow unresolved symbols in
libraries (but not executables). This is sometimes useful for
interposition but not in our case. Add a global linker flag to
ensure that all imported symbols can be resolved.

(Obviously, Linux and macOS linkers use a bit different syntax.)