Secure Cell: Seal mode
Introduce Secure Cell support starting with its sealing mode.

We keep the same structure as in JsThemis: the main module exports three
classes: SecureCellSeal, SecureCellTokenProtect, SecureCellContextImprint.
All of them have a constructor that accepts a master key, all of them
have "encrypt" and "decrypt" methods which accept a message and an
optional context.

Implementation involves quite a lot of copy-pasta so the individulal
modes have separate source files. Unfortunately, I was not able to
reduce code duplication much because of the subtle differences between
modes. Therefore the code is kept mostly dumb, involving various FFI
things we need to do to transfer the data into Emscripten heap and back.

Sealing mode is the easiest one and the most straightforward. Notable
things are processing of optional context arguments, overall interfacing
with C, and zeroing out of the sensitive master key after we're done
with it. Unfortunately, JavaScript does not have RAII and we have to
copy the data to Emscripten heap every time we need to call C code.