Secure Cell: Context Imprint mode
It's similar to the sealing mode, but the 'context' is required for
Context Imprint mode (duh...) We reflect this by producing a more
helpful error message if the user forgets to call encrypt/decrypt with
two arguments. The test set is also a bit different. And note that the
argument order for C functions is slightly different as well. Other that
that, it's mostly a copy of the sealing mode.