Secure Cell: Token Protect mode
Token Protect mode is different in that it returns *two* values: the
message and the authentication token separately. JavaScript does not
have tuples so we return an object with "data" and "token" fields,
just like JsThemis. For decryption we require data and token to be
provided as separate arguments. This is a bit weird for tests, but
should not cause issues in real-world usage.