Use "git describe" for versioning (after all)
We'd like to keep 'git describe' to make it easier for us to produce
development builds. However, we would still like to support git-less
builds.

Prefer the output of 'git describe' for versioning, but provide a more
simple fallback than we used before. Keep a VERSION file with version
which can be read to obtain the release version, if git is inaccessible.
Make sure to include this file into the distribution package.

This will give us versions like "0.11.1-42-ab430c" for builds from
untagged commits, "0.11.2" for builds directly from tagged versions,
or the content of the VERSION file if we're not building from a git
repository.

Note the "test -d", it makes sure that we don't accidentally pick up the
version of the parent project if Themis source code is vendored.