ABI-versioned library names (#454)
* Simplify package and library versioning

Current versioning approach in Makefile is way too complicated and does
not work for distributon tarballs when git repository is not available.
Fun fact: VERSION variable actually gets set twice.

Remove all this guesswork based on "git describe" and simply read the
version out of Makefile. Let it be the single source of truth.

* ABI-versioned library names

Change Soter and Themis libraries to contains ABI version information.
This will allow us to communicate future incompatible ABI changes to
their users and ensure that incompatible libraries are never loaded
accidentally.

Now that we have additional flags for linker, we can set macOS install
names directly rather than later with install_name_tool.

* Fix Linux packaging

We no longer create bogus symlinks for static libraris so we don't need
to install them. The symlinks are created during regular compilation so
we can drop the "symlink_realname_to_soname" target as redundant.

* Correct install names on macOS

Actually, with this changes the install names require more attention.
We still need to use "install_name_tool", but let's apply it only
during actual installation. That way we keep the libraries in build
directory unchanged, and we can still link against them for testing.
If we change the install name during build then the libraries become
unusable unless installed to the specified paths.

Pay attention to the paths we use (build path, install path, destdir
path), and the fact that we need to fixup Soter search path in Themis
library.

* Use "git describe" for versioning (after all)

We'd like to keep 'git describe' to make it easier for us to produce
development builds. However, we would still like to support git-less
builds.

Prefer the output of 'git describe' for versioning, but provide a more
simple fallback than we used before. Keep a VERSION file with version
which can be read to obtain the release version, if git is inaccessible.
Make sure to include this file into the distribution package.

This will give us versions like "0.11.1-42-ab430c" for builds from
untagged commits, "0.11.2" for builds directly from tagged versions,
or the content of the VERSION file if we're not building from a git
repository.

Note the "test -d", it makes sure that we don't accidentally pick up the
version of the parent project if Themis source code is vendored.