Really link Emscripten statically
While linking with "-lthemis" and *.so files seems to work at first,
that's not true for the case when Themis is already installed in the
system. Then emcc for some reason ignores the 'shared' Wasm libraries
in the build directory and tries linking against the native system
libraries (and failes).

Use the correct static linkage for Wasm targets: build the static libs,
and link against them instead of shared objects. However, continue
using shared objects for everything else.