Use dynamic linkage for tests (#460)
* Link tests dynamically: Soter

Switch to dynamic linkage for test binaries. This will allow us to
better track missing exported symbols and make sure that all public API
is actually tested.

The switch itself is more or less trivial: build the dynamic library
as the prerequisite, and link using '-Lbuild -lsoter' instead of adding
the static library to the object list. Since now the cryptographic
backend is managed by the dynamic library, we no longer have to link
against it explicitly.

Caveat for Linux: we need to set "rpath" to the build directory so that
the dynamic loader can locate Soter library there. Otherwise we'll have
to set LD_LIBRARY_PATH when running tests.

* Link tests dynamically: Themis

Link Themis Core tests dynamically, just like we do for Soter now.

Caveats:

  - the tests actually use some symbols from libsoter so we have
    to link against it as well, not only with just libthemis

  - setting LDFLAGS will leak them to libthemis and libsoter
    compilation by default, avoid that by using a separate variable

* Link tests dynamically: ThemisPP

Link ThemisPP tests dynamically, just like we do for Themis and Soter.
Overall it's just like with Themis.

* Disallow unresolved symbols during builds

The default behavior of linkers is to allow unresolved symbols in
libraries (but not executables). This is sometimes useful for
interposition but not in our case. Add a global linker flag to
ensure that all imported symbols can be resolved.

(Obviously, Linux and macOS linkers use a bit different syntax.)

* Link Emscripten tests statically

Unfortunately, Emscripten and Wasm do not support dynamic linkage (easily)
so they are configured for static linkage by default. This means that we
have to link in all the dependencies manually. Add cryptographic backend
to compilation flags so that it resolves correctly.

* Really link Emscripten statically

While linking with "-lthemis" and *.so files seems to work at first,
that's not true for the case when Themis is already installed in the
system. Then emcc for some reason ignores the 'shared' Wasm libraries
in the build directory and tries linking against the native system
libraries (and failes).

Use the correct static linkage for Wasm targets: build the static libs,
and link against them instead of shared objects. However, continue
using shared objects for everything else.