Fix Clippy warnings
New Clippy does not like when you write "|s| s.to_owned()" because
making a new closure here is reduntant (you can pass the function
std::borrow::ToOwned::to_owned there directly). This causes a warning
and we treat warnings as errors.

Anyway, it's not like we need to avoid this computation here with
a closure so rewrite the code to compute the default path in a more
straightforward way and thus avoid the wrath of our resident god of
static analysis.