WasmThemis: Secure Message support
Initial implementation of Secure Message wrapper. We export it as
three objects:

  - "SecureMessage" providing encrypt/decrypt mode

  - "SecureMessageSign" and "SecureMessageVerify" providing
    sign/verify mode

The reason for the split is to minimize possibility of API misuse.
Encryption and decryption is likely to be bidirectional, therefore
we require the users to specify both private and public keys.
(Though it might make sense to allow specifying "null" for either
of the keys in order to allow unidicational message flow.)
Signing and verifying signatures requires only one part of the key
pair, so these object accept a corresponding key.

Note that we allow *only* PrivateKey or PublicKey instances. This
ensures that we operate on valid keys every time.

As JavaScript is dynamically typed, we allow several constructor
syntax for SecureMessage. We can accept either a KeyPair (a single
entity), or two arguments with keys (in any order, so that devs
don't have to remember the correct one in absence of static typing
and helpful IDE hints).