Move "portable_endian.h" to Soter
We're going to need these functions in both Soter and Themis, so move
the header to Soter. Keep a stub in Themis to avoid breaking existing
code.

Also note that <sys/param.h> is a POSIX header and it is not available
on Windows. Remove it. All conversions are provided by <WinSock2.h>.