Turn Soter container accessors into functions
soter_container.h uses htonl() and ntohl() functions directly in macros.
This is not nice because it leaks implementation details and requires
POSIX API which is not available on Windows. Furthermore, this header
does not even include relevant headers for these functions!

Replace this macros with proper functions and move the implementation
to soter_container.c which includes "portable_endian.h". This allows
our API to not include private headers and does not unnecessarily expose
implementation details.

Theoretically, this will impact performance, but in practice the effect
should be negligible. These macros are not used in any hot paths, and
the compiler may even optimize these functions when using LTO.