Packaging for MSYS2 (and pacman)
Provide a PKGBUILD file which can be used to build Themis package for
MSYS2. A similar file could be used for building on ArchLinux. However,
this one does not work as is on Linux due to differences in dependency
naming and directory layout.

Themis is a library. It is customary in MSYS2 to provide Debian-style
split packages for libraries:

  - "themis" contains runtime shared libraries: *.dll on Windows
  - "themis-devel" contains headers, static libraries, and pkg-config
    files which are required for building software that uses Themis

The packages may be built by running

    makepkg -p PKGBUILD.MSYS2

in the directory with the PKGBUILD file, then you install them with

    pacman -U themis-0.11.2-1-x86_64.tar.xz
    pacman -U themis-devel-0.11.2-1-x86_64.tar.xz

makepkg always downloads pristine source tarball, so there is no point
in adding a Makefile target for preparing a package from source code.