Fix incorrect unsigned negation
MSVC produces an error for this line. It may be okay with gcc and clang,
but actually negating an unsigned value does not make sense, though it's
not clear what effect it should have according to the standard.

gcc and clang on x86_64 compile this:

    unsigned int b;
    b = -b;

into this:

    negl %eax

So they treat unsigned negation as two's complement, as if the number
was signed). Replace the current expression with its equivalent.

The algorithm implies that "b" may only have 0 and 1 as values. Thus
this line should be equivalent to

    b = b ? UINT_MAX : 0;

but who knows what weird code paths lead here, so let's play it safe
and use something completely equivalent. My little crypto: C is magic.