Windows support in JsThemis
Update binding.gyp file for Windows. This allows users to install
JsThemis on Windows provided that Themis Core is installed under
"C:\Program Files\Themis". We expect MSYS2-based build of Themis
Core to be installed there and use its naming convention and
directory layout. Unfortunately, Windows does not have a notion
of standard directories for development libraries and headers
so we have to tell MSBuild exactly where to look for the library.