MSYS2 target support (#469)
* MSYS2 target support

MSYS2 is a popular and easy way to compile POSIX-compliant software on
Windows. It provides GNU development environment and API compatibility
layer between POSIX and WinAPI. All of this allows us to use the same
build system and code as is, with the toolchain automagically producing
PE binaries useable on Windows.

However, we need to introduce a number of changes to the build system
to support MSYS:

- Detect MSYS environment and provide "IS_MSYS" variable for checks

  This allows us to introduce MSYS-specific behavior in Makefile.

- Build Windows DLLs for shared objects

  First of all, use MSYS naming scheme for dynamically loaded libraries:
  "msys-${library_name}-${abi_version}.dll". These will contain the code
  loaded by applications.

  After that we need to produce "import libraries" which are necessary
  in Windows world to link against dynamic libraries. On MSYS they are
  named "lib${library_name}.dll.a".

  Regular static libraries are still named "lib${library_name}.a". These
  are useable only for inside MSYS2.

- Install libraries to correct paths

  On MSYS shared libraries have to be installed to "/usr/bin" because of
  the way library loading works in Windows: libraries need to be placed
  next to executables.

  DLLs do not need symlinks on MSYS. However, they need import libraries
  instead. Update the installation targets to do this correctly.

  Update uninstallation targets as well.

- Avoid duplicate slashes when usign DESTDIR

  Remove slashes between DESTDIR and paths. Windows does not normalize
  duplicate slashes under MSYS which leads to really weird issues when
  creating directories.

- Place tests into the same directory as libraries

  We need to do this because tests link against libraries dynamically,
  they need to be able to load the library from the build directory
  (as it's not yet installed). Windows does not have a direct analog
  of RPATH (as on Linux), and does not record library paths (as macOS
  does). We can use LD_LIBRARY_PATH for running tests, but that would
  make them hard to use without our makefile. Therefore just change
  the location.

Also, the NIST test suite has a bit different name on Windows. Add it to
the .gitignore list.

With this, installation on MSYS can be done like this:

    make install

GGEZ!

(Note that MSYS does not have /usr/local in its default search paths.
You'll need to use -L/usr/local/lib when building applications, and
add /usr/local/bin to PATH when running them. Or just install with
PREFIX=/usr)

* Packaging for MSYS2 (and pacman)

Provide a PKGBUILD file which can be used to build Themis package for
MSYS2. A similar file could be used for building on ArchLinux. However,
this one does not work as is on Linux due to differences in dependency
naming and directory layout.

Themis is a library. It is customary in MSYS2 to provide Debian-style
split packages for libraries:

  - "themis" contains runtime shared libraries: *.dll on Windows
  - "themis-devel" contains headers, static libraries, and pkg-config
    files which are required for building software that uses Themis

The packages may be built by running

    makepkg -p PKGBUILD.MSYS2

in the directory with the PKGBUILD file, then you install them with

    pacman -U themis-0.11.2-1-x86_64.tar.xz
    pacman -U themis-devel-0.11.2-1-x86_64.tar.xz

makepkg always downloads pristine source tarball, so there is no point
in adding a Makefile target for preparing a package from source code.