Remove unnecessary macrology
All these fancy macros in soter_sign.c do not really make the code
easier to read and understand. They also do not reduce the amount
of copypasta (as it is evident for the diff stats). Replace them
with dumb code which provides better breakpoint experience and is
easier to understand.