Key length validation utilities
Add helper functions to Soter which validate the length of Soter
containers used for storing keys. There are slight differences
between RSA and EC keys due to the way they are laid out and
exported from cryptographic backends.

Soter uses portable structures to hide the implementation details of
cryptographic backends. This works quite well for RSA because it has
quite a few parameters and they are carefully put into the structure
one by one. Thus we use sizeof() to check their size.

However, for EC keys we have only one parameter and it's appended
directly to Soter container header. This leads to slight difference
between sizeof() of the structure and the length of actual data that
gets stored during export. Furthermore, we tell the users the actual
length, not sizeof the struct (which is kept private). Therefore for
EC keys we use their parameter length directly.

The new functions are privately exported API. We will not document
them, but they will be exported from shared objects for compatible
Themis to use.

Also note that we have to add the new files to Xcode project for them
to compile correctly when building Themis with Carthage.