Validate key sizes in Themis
Teach themis_is_valid_asym_key() to verify that keys have expected size
for their apparent kind, in addition to being in a correctly sized Soter
container. This adds additional resilience to using unrelated Soter
containers in place of keys (e.g., trying to pass a Secure Message
instead of a key).

We expect the keys to have correct size, not being truncated (which was
an error before) and not having additional bytes at the end (which *is*
an error from now on). Add tests to verify this.

Note that we do *not* validate the content of the keys for now. They
may still contain cryptographically nonsensical values. This will be
verified only when the keys are actually used to decrypt, sign, etc.
themis_is_valid_asym_key() may produce false positives.