Temporarily disable Android build on CircleCI
This job has been failing for literally *months*, keeping our build red
and preventing PRs from being merged cleanly. It is also quite slow in
comparison with other jobs.

Let's disable Android builds for now. The plan is it move Android CI
to Bitrise in the nearest time. There is no GitHub issue for this, but
the progress may be tracked internally at Cossack Labs [1].

[1]: https://ph.cossacklabs.com/T1133