Improve key validity checks (#486)
* Remove unnecessary macrology

All these fancy macros in soter_sign.c do not really make the code
easier to read and understand. They also do not reduce the amount
of copypasta (as it is evident for the diff stats). Replace them
with dumb code which provides better breakpoint experience and is
easier to understand.

* Key length validation utilities

Add helper functions to Soter which validate the length of Soter
containers used for storing keys. There are slight differences
between RSA and EC keys due to the way they are laid out and
exported from cryptographic backends.

Soter uses portable structures to hide the implementation details of
cryptographic backends. This works quite well for RSA because it has
quite a few parameters and they are carefully put into the structure
one by one. Thus we use sizeof() to check their size.

However, for EC keys we have only one parameter and it's appended
directly to Soter container header. This leads to slight difference
between sizeof() of the structure and the length of actual data that
gets stored during export. Furthermore, we tell the users the actual
length, not sizeof the struct (which is kept private). Therefore for
EC keys we use their parameter length directly.

The new functions are privately exported API. We will not document
them, but they will be exported from shared objects for compatible
Themis to use.

Also note that we have to add the new files to Xcode project for them
to compile correctly when building Themis with Carthage.

* Validate key sizes in Themis

Teach themis_is_valid_asym_key() to verify that keys have expected size
for their apparent kind, in addition to being in a correctly sized Soter
container. This adds additional resilience to using unrelated Soter
containers in place of keys (e.g., trying to pass a Secure Message
instead of a key).

We expect the keys to have correct size, not being truncated (which was
an error before) and not having additional bytes at the end (which *is*
an error from now on). Add tests to verify this.

Note that we do *not* validate the content of the keys for now. They
may still contain cryptographically nonsensical values. This will be
verified only when the keys are actually used to decrypt, sign, etc.
themis_is_valid_asym_key() may produce false positives.

* Preserve EC private key exporting quirks

It turned out that we have been using incorrect buffer size for private
keys all along. soter_engine_specific_to_ec_priv_key() should have used
ec_priv_key_size(), but it actually uses ec_pub_key_size() that returns
a buffer one byte bigger.

Welp, it's unfortunate, but this mistake is going to be set in stone and
will be declared a feature. Since we have been using public key length
for allocating private key buffers and such keys are already out there
in the wild, we cannot change the format now.

Use the public key length as expected length for exported private keys.
Add comments to the relevant exporting code to preserve its behavior.
Add a test to verify that exported private key has the same length as
the public one and that it has expected padding.

* Constants for key size tags

Introduce helper character constants for key size tags. Use them instead
of hardcoding '2' everywhere. We have the same string constants already,
but they cannot be used in switch expressions so we have to introduce
new constants for that.